+++ tests/testthat/test-plot_grpfrq.R
@@ -1,5 +1,7 @@
 if (suppressWarnings(
   require("testthat") &&
+  require("sjlabelled") &&
+  require("haven") &&
   require("sjPlot")
 )) {